requirements.txt
# Dados de mercado
yfinance>=0.2.28

# Processamento de dados
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning e Explicabilidade
scikit-learn>=1.3.0

# LangChain - Framework para LLMs
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0

# HTTP requests (para Ollama e scraping)
requests>=2.31.0
beautifulsoup4>=4.12.0

# Utilitários
python-dotenv>=1.0.0
